Set the ttl again when the job is being performed.
The idea is that for a system with very large queues (millions of jobs) a job may sit in the queue for a long time (a few days), but once it has started to be performed, it should be performed quickly (say, 30 minutes).  Thus, we want to shorten the TTL at that point.